# Improved Encoder API