This PR reworks the implementation of the `encoder` module to accomplish 2 goals:

1. Improve the performance of the `encoder` logic. The existing implementation performs a *lot* of small allocations due to creating and passing around around small `Vec` and `String` objects internally. The reworked implementation passes through a reference to the output, and all data is written directly to the output; no internal buffers are allocated, making performance much faster. The performance improvement from this is significant, improving performance by 4-40x in my results from stress-test benchmarks that I created alongside these changes (sample results below). In a practical project in which I'm using `rosc`, these changes have decreased the amount of time spent encoding OSC packets to about 1/8 of what it was before (from ~150 us per packet to ~18 us) with no code changes in that project.

2. Make the output more flexible. This PR adds an `Output` trait which allows the user control over what is done with the output. For `std`, this includes an implementation of this trait for `std::io::Write`, allowing OSC data to be encoded directly to a `Vec<u8>`, `Cursor`, socket, file, etc. In the non-`std` case, an implementation is provided only for `Vec<u8>`. These changes also include `encode_into` and `encode_string_into` functions, which allow the user to pass in a mutable reference to any object implementing this `Output` trait, allowing the user to control how the encoded data is handled.

The existing public functions have not had their signatures altered, only new public items have been added. Thus, according to the [SemVer Compatibility guide](https://doc.rust-lang.org/cargo/reference/semver.html), these changes should only qualify as a minor change. Additionally, these changes do not incur a MSRV bump (1.52 is the minimum supported version according to `cargo-msrv`, both with and without these changes). Thus, these changes don't require a major version bump.

This also fixes unit tests not building without the `std` feature. All existing unit tests are passing, both with and without `std`. No additional unit tests have been added.

Benchmark results on my machine for comparison:

| Benchmark | Before | After | Relative Speed (1x = no change) |
|--------------|--------|-------|----------------|
| bench_encode_args_bool | 35,464 ns/iter ± 820 | 8,464 ns/iter ± 149 | 4.2x |
| bench_encode_args_double | 70,147 ns/iter ± 1,944 | 9,284 ns/iter ± 154 | 7.6x |
| bench_encode_args_float | 69,832 ns/iter ± 3,266 | 9,053 ns/iter ± 233 | 7.7x |
| bench_encode_args_int | 69,735 ns/iter ± 2,531 | 9,031 ns/iter ± 190 | 7.7x |
| bench_encode_args_long | 69,747 ns/iter ± 916 | 9,301 ns/iter ± 165 | 7.5x |
| bench_encode_args_nil | 35,013 ns/iter ± 1,173 | 8,201 ns/iter ± 137 | 4.3x |
| bench_encode_args_string | 143,910 ns/iter ± 2,152 | 13,646 ns/iter ± 151 | 10.5x |
| bench_encode_bundles | 294,400 ns/iter ± 23,868 | 7,092 ns/iter ± 102 | 41.5x |
| bench_encode_bundles_into | - | 7,389 ns/iter ± 306 | - |
| bench_encode_messages | 374,490 ns/iter ± 3,665 | 12,190 ns/iter ± 116 | 30.7x |
| bench_encode_messages_into | - | 14,675 ns/iter ± 149 | - |