@klingtnet 2023-03-09
@Barinzaya That's awesome work you did there 🐎 !

I'll try to find some time in the next days, maybe next week, to get a deeper look into the changes and give it a proper review. Thanks in advance!

Edit: I deleted the original comment, because I accidentally made that using my work account.