@Barinzaya 2023-03-09
I simplified the `encode_args` benchmarks to simplify the packets being encoded and focus it more on the arguments (was `bundle(1 message(args))`, now just `message(args))`. For some reason, this seems to have improved the performance on the branch further (expected, since there are fewer steps to the encoding), but for some reason didn't seem to have any effect on the benchmark time on `master`, much to my surprise. This makes the improvement numbers look nicer, but also makes me wonder if the benchmarks I added are missing some case that accounts for that difference.

Just wanted to make an open note of it, since it gives me the feeling of fluffing the numbers, which wasn't the intention at all.