@klingtnet 2023-03-11
@Barinzaya Can you share the command (if there is some) that you used to create the benchmark table from https://github.com/klingtnet/rosc/pull/44#issue-1617160005 ?

Did you use https://github.com/BurntSushi/cargo-benchcmp ?