@Barinzaya 2023-03-11
As far as the benchmark comparison table, I created them manually. Ran `cargo bench` on `master` and copied the output, then switched to `encoder-improvements` and copied its outputs. The relative speeds were also done manually (old time / new time).

I probably could have at least used a spreadsheet or something, but there were only a few benchmarks when I started and I never bothered to set one up. There might be other, more suitable tools out there too.