@Barinzaya 2023-03-29
Commits have been squashed.

Unsquashed commits are still available in the `encoder-improvements+unsquashed` branch on the fork's repository.