@Barinzaya 2023-04-01
Added one more commit, which adds an encoder unit test using `encode_into` to write to a Vec via a Cursor, using the `WriteOutput` struct. I feel this is somewhat important, as `Output` implementation for `WriteOutput` is not covered by the existing unit tests.

That makes this 5 commits, one per file touched. I can squash them down a bit further (e.g. combine the commit for the unit test change with the commit for the encoder changes) if preferred.